backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
google-generativeai==0.8.4
python-dotenv==1.0.1
python-multipart==0.0.9
datasets>=2.18.0
tqdm>=4.66.0
pandas>=2.0.0